[refactor/#224] shortDropdown 폴더명 PascalCase로 변경, import 경로 수정
